tooltip: Change placement of tooltip according to the popover menu.
This commits changes the placement of "Add streams" tooltip
and "Filter streams" tooltip to "bottom" when the
"Add streams" popover menu is opened and changes its back
to "top" when the popover menu is closed.

It makes use of the "id" attribute that has been assigned
to those tooltips in commit 01e6121e5a.

Fixes: #20675.