node-tests: Remove mock of `topic_list` in `topic_list_data.test`.
With commit f0d1332ba2f4, we no longer need to mock `topic_list.js`
in the `topic_list_data` node test. Instead, we can pass a string
directly to that function. The default is an empty string unless
there is a search input. Also, updates the default of the zoom to
be false instead of undefined in the tests.