todo_widget: Fix cursor focus in todo lists.
Previously, when adding a task to a /todo list, after pressing "Add
task", the cursor would move to the "Description" field, which is
not the desired behavior. This commit fixes this issue by ensuring
that the cursor moves to the "New task" field after a task is created.

The core bug was introduced in
9a6ac086bbc08c6662f38c85428805323f73e696, which added the incorrect
.focus().

Also tighten the selectors while we're at it.

Fixes #25064.