typeahead: Do not prioritize exact name matches for users.
As discussed in the comment, this resulted in the many inactive users
in chat.zulip.org with "Tim" as their full display name from being
incorrectly prioritized over me in a stream where I've been active;
and this class of problem seems like it will be common in large open
communities.

This reverts a portion of dc61efaae268f256112f9708939a0e00439e2711,
with explanatory comments for why this behavior is preferred.

The stream name changes in the original PR remain very helpful.