unread: Fix exception handling mentions in private messages.
Private messages with mentions are included in
unread_mentions_counter, but of course don't have a stream ID, so
should be skipped when calculating which streams contain unread
mentions.

This code has always been wrong, it put `undefined` in the set of
stream IDs with unread mentions.

What changed recently is that in
c9ae83037c8adc223c4a246f440c9d27a2fa3cc7, we started using the
`stream_id` to do an additional lookup in the unmuted version of the
function, and doing that lookup with `undefined` threw an exception.