bootstrap: Remove unused stylings for `btn-info`.
Currently, we are in the process of removing bootstrap out of the current
Zulip codebase. A quick git grep on btn-info shows that the class is only
used in `templates/corporate/billing.html`. But it doesn't take advantage
or use most of the styling rules that are set.

We should get rid of those rules as it's not being used and help simplify
the process in removing bootstrap.