api-docs: Add when moving messages fields/parameters were added.
Adds API changelog feature level 1 and associated Changes notes
for when the `stream_id` parameter in the `PATCH /messages/message_id`
was added, and for when the `prev_stream` field was added to edit
history information for messages.

We're adding these to the Zulip 3.0 feature level 1 because
commit 2d753b5aaf95e5c that introduced this field and this parameter
to the server / backend code was merged before the commit that added
the API feature level tracking, commit 2cecefc1b2345f, at level 1.