search: Stop disabling and enabling the search close button.
This is logic from 10 years ago (dbc4798594c4aa859013e46fcbe94977fd43cdf4)
that is no longer relevant. It seems like we used to show the
search bar open all the time and only showed the buttons when
there was focus in the search bar. Now we close the search bar
when it's not being used, and no longer need to update button
visibility or disable the search close button.