api-docs: Correct example for realm_bot delete event.
Updates the example for the realm_bot delete event so that it does
not have a full_name field.

This was a pre-existing error in the documentation when the remove
and delete events shared the same event documentation. They were
separated in the documentation in commit bbcda261c4ffc42.

The difference between these two events was noted when they were
added to `event_schema.py` in commit 9c0639683f4559.