slack: Handle the special case of permissions denied on team.info call.
This is a follow-up to 869d634926a7feffc956f68e57c34be72c590cd9, for
the case when the `team:read` permission is missing, which causes the
`team.info` call itself to fail.  The error message supplies
information about the provided and missing permissions -- but it also
still sends the `X-OAuth-Scopes` header which we normall read, so we can
use that as normal.