test_unread_msgs: Update comment regarding unread message count.
A comment was added in 14a8d3a to convey that the unread count
at that time doesn't exclude the unreads in muted topics.

cc2ad07 added the support to exclude the muted topic;
however, the comment was not updated.

This commit updates the comment to reflect the current behavior.