realm: Add uuid and associated secret columns.
Thisi and the following commit follow the approach used in
6b7bbdd9853fdfb4cdb0ce2bbece1a684cfe88c6.

First migration requires a server restart - after that any new realms
will be created with the columns set.

The following migrations are in the next commit:

Second migration does a backfill for older realms and can run in the
background while the server is operating normally.

Third migration enforces null=False now that all realms have the columns
set.