register: Update placeholder for name input.
This commit updates the placeholder for name input in account
creation form to "Your full name".

We also removes the "名前" since that confuses the user who
do not know the language. This text was originally added to
indicate that the name field supports unicode characters, which
most modern apps do currently, so there is no need to add this
text in the placeholder.

Fixes part of #29226.