compose_state: Remove incorrect recipient_has_topics helper.
This function was introduced in
c1f972f30ea4df25060d1632a7dbf26f133c54d6 and always has the opposite
value from what its name suggests. It worked OK because the caller
also expected it to have an incorrect value.

It's cleaner in this context to just use the `stream_id` helper, so
just delete the bad code and use that.

Fix the docstring comment having been incorrectly pushed down while
we're addressing this.